user based requests

+++ src/routes/apiRoutes.ts
@@ -7,7 +7,7 @@ const router = express.Router()
 // get All todo List
 router.get('/all', async (req: Request, res: Response) => {
     try{
-        let todos: Todo[] = await db.Todo.findAll()
+        let todos: Todo[] = await db.Todo.findAll({where: {userId: req.body.userObj.id}})
         if(todos === null){
             return res.status(204).send({msg: 'No Records Found'}) // No content
         }
@@ -20,7 +20,7 @@ router.get('/all', async (req: Request, res: Response) => {
 // get By ID
 router.get('/find/:id', async (req: Request, res: Response) => {
     try{
-        let todo: Todo = await db.Todo.findOne({ where: {id: req.params.id }})
+        let todo: Todo = await db.Todo.findOne({ where: {id: req.params.id, userId: req.body.userObj.id }})
         if(todo === null) res.status(204).send({msg: 'Todo Not Found'}) // No content
         else res.status(200).send({data: todo}) // Sent data
     }catch(err){
@@ -31,7 +31,7 @@ router.get('/find/:id', async (req: Request, res: Response) => {
 // Inserting new Route
 router.post('/new', async (req: Request, res: Response) => {
     try{
-        let todo: Todo = await db.Todo.create({ text: req.body.text })
+        let todo: Todo = await db.Todo.create({ text: req.body.text, userId: req.body.userObj.id })
         if(todo !== null) res.status(200).send({msg: "Todo Added Successfully"}) // Data sent
         else res.status(400).send({msg: 'Todo not created.'}) // Bad request
     }catch(err){
@@ -42,7 +42,7 @@ router.post('/new', async (req: Request, res: Response) => {
 // delete todo
 router.delete('/delete/:id', async (req: Request, res: Response) => {
     try {
-        await db.Todo.destroy({ where: {id: req.params.id }})
+        await db.Todo.destroy({ where: {id: req.params.id, userId: req.body.userObj.id}})
         res.status(200).send({msg: 'Successfully deleted'})
     }catch(err) {
         res.status(503).send({msg: `Somethong wen't wrong : ${err}`}) // Service Unavaliable
@@ -52,7 +52,7 @@ router.delete('/delete/:id', async (req: Request, res: Response) => {
 // Update 
 router.put('/edit', async (req: Request, res: Response) => {
     try {
-        await db.Todo.update({text: req.body.text}, {where: {id: req.body.id }})
+        await db.Todo.update({text: req.body.text}, {where: {id: req.body.id, userId: req.body.userObj.id}})
         res.status(200).send({msg: 'successfully Edited'})
     }catch(err){
         res.status(503).send({msg: `Somethong wen't wrong : ${err}`}) // Service Unavaliable
